+++ cmd/leaf/main.go
@@ -1,0 +1,6 @@
+package main
+
+// start the leader
+
+// handle configuration
+// scrape intervals, batching of requests to the scheduler, etc.
