+++ functions/go/hello/main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 
-	functionRuntimeInterface "github.com/3s-rg-codes/HyperFaaS/pkg/functionRuntimeInterface"
+	functionRuntimeInterface "github.com/3s-rg-codes/HyperFaaS/pkg/worker/functionRuntimeInterface"
 )
 
 func main() {
