+++ functions/go/sleep/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	functionRuntimeInterface "github.com/3s-rg-codes/HyperFaaS/pkg/functionRuntimeInterface"
+	functionRuntimeInterface "github.com/3s-rg-codes/HyperFaaS/pkg/worker/functionRuntimeInterface"
 )
 
 func main() {
