+++ pkg/leaf/api/api.go
@@ -1,0 +1,10 @@
+package api
+
+// grpc api endpoints that leafLeader will expose
+
+// these will be called by Leaders above in the leader chain
+
+// potentially also some endpoints could be called by workers ??
+
+// The main "state" of the leader is all of the worker IPs.
+// Would only be changed if a worker is added or removed.
